# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NumShield API
  version: 1.0.0
  description: Cameroon phone-number normalization, validation, and operator detection.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080
paths:
  /api/v1/number/verify:
    post:
      summary: Verify and enrich a Cameroon phone number
      operationId: verify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneNumberRequest'
      responses:
        '200':
          description: Verification completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/phone-numbers/normalize:
    get:
      summary: Normalize a Cameroon phone number
      operationId: normalizeGet
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Number normalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizationApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Normalize a Cameroon phone number
      operationId: normalizePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneNumberRequest'
      responses:
        '200':
          description: Number normalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizationApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/phone-numbers/validate:
    get:
      summary: Validate a Cameroon phone number
      operationId: validateGet
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Number validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Validate a Cameroon phone number
      operationId: validatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneNumberRequest'
      responses:
        '200':
          description: Number validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  parameters:
    Number:
      name: number
      in: query
      required: true
      schema:
        type: string
      example: '690123456'
  responses:
    BadRequest:
      description: Invalid request or phone number
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
    PhoneNumberRequest:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
          minLength: 1
          example: '690123456'
    VerificationApiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/VerificationResponse'
    NormalizationApiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/NormalizationResponse'
    ValidationApiResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ValidationResponse'
    SuccessEnvelope:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        timestamp:
          type: string
          format: date-time
    ErrorApiResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ApiError'
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [code, message, stage]
      properties:
        code:
          type: string
          enum: [INVALID_REQUEST, INVALID_PHONE_NUMBER]
        message:
          type: string
        stage:
          type: string
          enum: [REQUEST_VALIDATION, NORMALIZATION, VALIDATION]
    VerificationResponse:
      type: object
      required: [input, normalized, valid, operator, countryCode, country]
      properties:
        input:
          type: string
        normalized:
          type: string
          example: '+237690123456'
        valid:
          type: boolean
          const: true
        operator:
          type: string
          enum: [MTN, ORANGE, NEXTTEL, CAMTEL, UNKNOWN]
        countryCode:
          type: string
          const: '237'
        country:
          type: string
          const: CM
    NormalizationResponse:
      type: object
      required: [raw, normalized]
      properties:
        raw:
          type: string
        normalized:
          type: string
          example: '+237690123456'
    ValidationResponse:
      type: object
      required: [phoneNumber, valid]
      properties:
        phoneNumber:
          type: string
          example: '+237690123456'
        valid:
          type: boolean
          const: true